Basic Hangman Logic (Not Finished)
I have inserted a 2D array of sorts, to present the image of a hangman. I have scripted input and verification of input compared to alphabet array. 

Work In-Progress...
Blank character slots for word in play
If input matches a letter in the word in play, blank space is filled with the input letter.