Added Blank Character Slots
blank slots now show up, and if you guess the correct letter it fills in the first slot with this letter. However in doing this it writes over the current document and deletes everything else. To battle this I will have to read and write to the html file directly every time the gameplay is updated.