Core Game Complete
I finished off the reactive word so that the user inputs, if correct, are splice"d"  to the reactive word. If a guess is wrong, the lives go down, and the Image is updated. If all lives are lost the user is prompted with a GAME OVER alert. If they win they are prompted with a WIN alert.

When a guess is made the letters array is updated, as the guess is removed from the array.  If the guess is wrong the guess value is moved into a array called sinBin, which holds all the wrong guesses.