Cleaned Up Code
removed all the console.log() calls as they were how I debugged a lot of my code using the console log.

I also changed the <p>Used Letters:</p> to <p>Wrong Guesses:</p>, as any correct guesses are not put in the sinBin, which is where wrong guesses are stored, so this removes any confusion.

Comment Update:
You can no longer check via inspect element, as in the past the word to guess was logged in the console due to debugging. ;)